pyproject.toml
[tool.poetry]
name = "drftest"
version = "0.1.0"
description = ""
authors = ["Sylvia47 <via1100302047@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
Django = "^3.2.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
